requirements.txt
openai
langchain
langchain_openai
langchain_pinecone
langchain_experimental
langchain_core
langchain_community
langsmith
python-dotenv
langchainhub
bs4
faiss-cpu
pypdf